requirements.txt
python-dotenv~=0.15.0
pytest~=6.2.1
dnstwist~=20200707
asyncpg~=0.21.0
aiohttp~=3.7.3
requests~=2.25.1
beautifulsoup4~=4.9.1
sqlalchemy~=1.4.0b3
python-whois~=0.7.3
google~=3.0.0